Day 17 Part 2
Day 17 Part 2 Completed

When over two rows, all x values are blocked, the rows below these are removed from memory to save space.
Each time this happens the 'game-state' is saved. If a game-state if found to be repeating, this means it will repeat in a loop forever and the calculations do not need to be repeated.